fix: replace persist by firebase auth listener
