+++ src/features/private-route/index.tsx
@@ -1,15 +1,13 @@
-import { Navigate, useLocation } from 'react-router-dom';
+import { Navigate } from 'react-router-dom';
 
 import { useAuth } from '@/entities/user';
 import { TypeAppRoute } from '@/shared/config';
 
 export const PrivateRoute = ({ children }: IPrivateRouteProps): JSX.Element => {
-  const location = useLocation();
-
   const { isAuth } = useAuth();
 
   if (!isAuth) {
-    return <Navigate to={TypeAppRoute.Login} state={{ from: location }} />;
+    return <Navigate to={TypeAppRoute.Welcome} replace />;
   }
 
   return children;
